Move gnocchi.storage.common to gnocchi.common
