+++ gnocchi/common/ceph.py
