+++ gnocchi/common/redis.py
