+++ gnocchi/common/s3.py
