+++ gnocchi/common/swift.py
